Updates to Result Message

+++ src/simulation/util/Cell.java
@@ -22,7 +22,7 @@ public interface Cell<T> {
 
 	public double getTemp();
 	
-	public void setTemp(float temp);
+	public void setTemp(double temp);
 	
 	public void swapTemp();
